refactor: rename MapSVGProps to VsmData and outsource to a types.ts
